updated to ES 5.0.0 RC1 (deguice completed for now)
