+++ plugin-descriptor.properties
@@ -3,7 +3,7 @@
 description=Provides SSL for Elasticsearch 5
 #
 # 'version': plugin's version
-version=5.0.0-beta1-16a
+version=5.0.0-rc1-16a
 #
 # 'name': the plugin name
 name=search-guard-ssl
@@ -22,4 +22,4 @@ java.version=1.8
 # elasticsearch release. This version is checked when the plugin
 # is loaded so Elasticsearch will refuse to start in the presence of
 # plugins with the incorrect elasticsearch.version.
-elasticsearch.version=5.0.0-beta1
+elasticsearch.version=5.0.0-rc1
