+++ pom.xml
@@ -16,7 +16,7 @@
 
     <groupId>com.floragunn</groupId>
     <artifactId>search-guard-ssl</artifactId>
-    <version>5.0.0-beta1-16c-SNAPSHOT</version>
+    <version>5.0.0-rc1-16c-SNAPSHOT</version>
     <packaging>jar</packaging>
 
     <name>Search Guard SSL</name>
@@ -44,7 +44,7 @@
     </licenses>
 
     <properties>
-        <elasticsearch.version>5.0.0-beta1</elasticsearch.version>
+        <elasticsearch.version>5.0.0-rc1</elasticsearch.version>
         <netty-native.version>1.1.33.Fork23</netty-native.version>
         <dependency.locations.enabled>false</dependency.locations.enabled>
     </properties>
@@ -104,7 +104,7 @@
         <dependency>
             <groupId>org.elasticsearch.plugin</groupId>
             <artifactId>transport-netty4-client</artifactId>
-            <version>5.0.0-beta1</version>
+            <version>${elasticsearch.version}</version>
         </dependency>
 
         <!-- Only test scoped dependencies hereafter -->
